cloud: parse and store cloud in settings
Cloud is determined from config and stays the same from that moment,
so it makes sense to parse it once and store in settings, so whoever
needs it, could just use self.settings.cloud. This is going to be
more useful in the few next commits, where push/pull/status are introduced.

Signed-off-by: Ruslan Kuprieiev <kupruser@gmail.com>