Run pytype type checker as part of CI
* Run pytype as part of tests on Ubuntu

Pytype is currently not supported at all on Windows.  Since pytype's results
_should_ be completely platform-agnostic and Linux is its primary target OS,
it's not worth slowing down macOS builds to re-run it there.

It _is_ worth running pytype as a test (as opposed to lint) to make sure the
types are valid in all of the supported versions of Python since later versions
introduced additional types to the stdlib.

Addresses #4697.