push: prevent crash while using sqlite db on windows network share
There is no strict convention for URI paths that refer to UNC shares on
windows. Although the current model is using the most common approach,
this is not the approach that sqlite expects, and a authority that is
blank or "localhost" is required. This prevents a crash when pushing
a DVC repo that is hosted on a windows network share affected by
the change in python 3.8 that resolves network shares to their UNC
path when using os.path.abspath. Note that drive letter paths on windows
are handled identically to how they were beforehand.